Add navigation bar, add separate directories for components and styles
